requirements/prod.txt
pydantic==2.3.0
newsapi-python==0.2.7
pydantic-settings==2.0.3
